Milestone 9: Configuration & Security Hardening (4 steps)
- ConfigValidationStep: Validate Pydantic settings, compare .env/.env.example
- SecretsDetectionStep: Enhanced secrets detection with entropy analysis
- EnvCompletenessStep: Verify env var documentation completeness
- SecurityHeadersStep: Detect Flask/FastAPI/Django security headers

Output files:
- logs/120_config_validation.txt
- logs/121_secrets_advanced.txt
- meta/120_config_completeness.txt
- logs/122_security_headers.txt

Integration:
- Updated profiles.py with step imports and conditional inclusion
- Updated context.py with no_config_security_analysis option
- Updated cli.py with --config-security-analysis flag
- All steps tested and verified working

+++ pybundle/steps/container_image.py
@@ -284,7 +284,7 @@ class ContainerImageStep(Step):
 
         return layers
 
-    def _format_size(self, size_bytes: int) -> str:
+    def _format_size(self, size_bytes: float) -> str:
         """Format bytes into human readable size."""
         for unit in ["B", "KB", "MB", "GB", "TB"]:
             if size_bytes < 1024:
